- changed CRLF to LF - fixed warnings - added 64bit support on Windows

+++ fastgrid_master/fastgrid/openthreads/win32_src/Win32ThreadPrivateData.h
@@ -47,7 +47,7 @@ private:
     Win32ThreadPrivateData() {};
     ~Win32ThreadPrivateData();
 
-    unsigned int stackSize;
+    size_t stackSize;
     bool stackSizeLocked;
     bool isRunning;
